+++ src/bases/BaseValidator.ts
@@ -1,3 +1,8 @@
+import vine from "@vinejs/vine";
 import "@/utils/vine";
 
-export default class BaseValidator {}
+export default class BaseValidator {
+    public static get validator(): typeof vine {
+        return vine;
+    }
+}