+++ CHANGELOG.md
@@ -3,6 +3,27 @@ All notable changes to this project will be documented in this file.
 
 ---
 
+## [v0.1.23](https://github.com/crenata/bejibun-core/compare/v0.1.18...v0.1.23) - 2025-10-14
+
+### 🩹 Fixes
+
+### 📖 Changes
+New Features :
+- Base class (Controller, Model, Validator)
+- Bootstrap for startup load
+- Adding some utils (Chalk, Response, Router, Soft Deletes)
+- Adding some types (Middleware, Router, Validator, etc)
+- Vine validation macro for `exists` and `unique`
+- Adding validation exception
+
+### ❤️Contributors
+- Havea Crenata ([@crenata](https://github.com/crenata))
+- Ghulje ([@ghulje](https://github.com/ghulje))
+
+**Full Changelog**: https://github.com/crenata/bejibun-core/blob/master/CHANGELOG.md
+
+---
+
 ## [v0.1.18](https://github.com/crenata/bejibun-core/compare/v0.1.0...v0.1.18) - 2025-10-12
 
 ### 🩹 Fixes